Push partition spec into physical plan.
