+++ daft/daft.pyi
@@ -805,7 +805,7 @@ class LogicalPlanBuilder:
 
     @staticmethod
     def in_memory_scan(
-        partition_key: str, cache_entry: PartitionCacheEntry, schema: PySchema, partition_spec: PartitionSpec
+        partition_key: str, cache_entry: PartitionCacheEntry, schema: PySchema, num_partitions: int
     ) -> LogicalPlanBuilder: ...
     @staticmethod
     def table_scan(
@@ -817,7 +817,10 @@ class LogicalPlanBuilder:
     def explode(self, to_explode: list[PyExpr]) -> LogicalPlanBuilder: ...
     def sort(self, sort_by: list[PyExpr], descending: list[bool]) -> LogicalPlanBuilder: ...
     def repartition(
-        self, num_partitions: int, partition_by: list[PyExpr], scheme: PartitionScheme
+        self,
+        partition_by: list[PyExpr],
+        scheme: PartitionScheme,
+        num_partitions: int | None,
     ) -> LogicalPlanBuilder: ...
     def coalesce(self, num_partitions: int) -> LogicalPlanBuilder: ...
     def distinct(self) -> LogicalPlanBuilder: ...
@@ -834,7 +837,6 @@ class LogicalPlanBuilder:
         compression: str | None = None,
     ) -> LogicalPlanBuilder: ...
     def schema(self) -> PySchema: ...
-    def partition_spec(self) -> PartitionSpec: ...
     def optimize(self) -> LogicalPlanBuilder: ...
     def to_physical_plan_scheduler(self) -> PhysicalPlanScheduler: ...
     def repr_ascii(self, simple: bool) -> str: ...
