+++ daft/dataframe/dataframe.py
@@ -24,13 +24,7 @@ from typing import (
 from daft.api_annotations import DataframePublicAPI
 from daft.context import get_context
 from daft.convert import InputListType
-from daft.daft import (
-    FileFormat,
-    JoinType,
-    PartitionScheme,
-    PartitionSpec,
-    ResourceRequest,
-)
+from daft.daft import FileFormat, JoinType, PartitionScheme, ResourceRequest
 from daft.dataframe.preview import DataFramePreview
 from daft.datatype import DataType
 from daft.errors import ExpressionTypeError
@@ -89,8 +83,11 @@ class DataFrame:
         if self._result_cache is None:
             return self.__builder
         else:
+            num_partitions = self._result_cache.num_partitions()
+            # Partition set should always be set on cache entry.
+            assert num_partitions is not None
             return self.__builder.from_in_memory_scan(
-                self._result_cache, self.__builder.schema(), self.__builder.partition_spec()
+                self._result_cache, self.__builder.schema(), num_partitions=num_partitions
             )
 
     def _get_current_builder(self) -> LogicalPlanBuilder:
@@ -125,9 +122,6 @@ class DataFrame:
         if show_optimized:
             builder = builder.optimize()
         print(builder.pretty_print(simple))
-
-    def num_partitions(self) -> int:
-        return self.__builder.num_partitions()
 
     @DataframePublicAPI
     def schema(self) -> Schema:
@@ -306,7 +300,7 @@ class DataFrame:
 
         context = get_context()
         cache_entry = context.runner().put_partition_set_into_cache(result_pset)
-        builder = LogicalPlanBuilder.from_in_memory_scan(cache_entry, parts[0].schema())
+        builder = LogicalPlanBuilder.from_in_memory_scan(cache_entry, parts[0].schema(), result_pset.num_partitions())
         return cls(builder)
 
     ###
@@ -345,7 +339,7 @@ class DataFrame:
             cols = self.__column_input_to_expression(tuple(partition_cols))
             for c in cols:
                 assert c._is_column(), "we cant support non Column Expressions for partition writing"
-            self.repartition(self.num_partitions(), *cols)
+            self.repartition(None, *cols)
         else:
             pass
         builder = self._builder.write_tabular(
@@ -386,7 +380,7 @@ class DataFrame:
             cols = self.__column_input_to_expression(tuple(partition_cols))
             for c in cols:
                 assert c._is_column(), "we cant support non Column Expressions for partition writing"
-            self.repartition(self.num_partitions(), *cols)
+            self.repartition(None, *cols)
         else:
             pass
         builder = self._builder.write_tabular(
@@ -614,7 +608,7 @@ class DataFrame:
         return count_df.to_pydict()["count"][0]
 
     @DataframePublicAPI
-    def repartition(self, num: int, *partition_by: ColumnInputType) -> "DataFrame":
+    def repartition(self, num: Optional[int], *partition_by: ColumnInputType) -> "DataFrame":
         """Repartitions DataFrame to ``num`` partitions
 
         If columns are passed in, then DataFrame will be repartitioned by those, otherwise
@@ -625,8 +619,8 @@ class DataFrame:
             >>> part_by_df = df.repartition(4, 'x', col('y') + 1)
 
         Args:
-            num (int): number of target partitions.
-            *partition_by (Union[str, Expression]): optional columns to partition by.
+            num (Optional[int]): Number of target partitions; if None, the number of partitions will not be changed.
+            *partition_by (Union[str, Expression]): Optional columns to partition by.
 
         Returns:
             DataFrame: Repartitioned DataFrame.
@@ -1164,9 +1158,7 @@ class DataFrame:
         partition_set, schema = ray_runner_io.partition_set_from_ray_dataset(ds)
         cache_entry = context.runner().put_partition_set_into_cache(partition_set)
         builder = LogicalPlanBuilder.from_in_memory_scan(
-            cache_entry,
-            schema=schema,
-            partition_spec=PartitionSpec(PartitionScheme.Unknown, partition_set.num_partitions()),
+            cache_entry, schema=schema, num_partitions=partition_set.num_partitions()
         )
         return cls(builder)
 
@@ -1233,9 +1225,7 @@ class DataFrame:
         partition_set, schema = ray_runner_io.partition_set_from_dask_dataframe(ddf)
         cache_entry = context.runner().put_partition_set_into_cache(partition_set)
         builder = LogicalPlanBuilder.from_in_memory_scan(
-            cache_entry,
-            schema=schema,
-            partition_spec=PartitionSpec(PartitionScheme.Unknown, partition_set.num_partitions()),
+            cache_entry, schema=schema, num_partitions=partition_set.num_partitions()
         )
         return cls(builder)
