+++ src/daft-plan/src/optimization/rules/mod.rs
@@ -1,11 +1,11 @@
-mod drop_repartition;
+// mod drop_repartition;
 mod push_down_filter;
 mod push_down_limit;
 mod push_down_projection;
 mod rule;
 mod utils;
 
-pub use drop_repartition::DropRepartition;
+// pub use drop_repartition::DropRepartition;
 pub use push_down_filter::PushDownFilter;
 pub use push_down_limit::PushDownLimit;
 pub use push_down_projection::PushDownProjection;
