+++ src/daft-plan/src/optimization/rules/push_down_limit.rs
@@ -196,7 +196,7 @@ mod tests {
             Field::new("a", DataType::Int64),
             Field::new("b", DataType::Utf8),
         ])
-        .repartition(1, vec![col("a")], PartitionScheme::Hash)?
+        .repartition(Some(1), vec![col("a")], PartitionScheme::Hash)?
         .limit(5, false)?
         .build();
         let expected = "\
